// frontend/frontend/src/types.ts
export type IdParam = string | undefined | number;
// ...
export type ImageType = 'EVENT_COVER' | 'EDITOR_IMAGE';
// ...
export interface EventBase {
    title: string;
    description?: string;
    start_date: string;
    end_date?: string;
    hostId:string;
}
// types.ts
export type Event = {
  id: string;
  title: string;
  description: string;
  eventDate: string; // ISO date string
  location: string;
  validated: boolean;
  host: {
    id: string;
    name: string;
  };
  category?: {
    id: string;
    name: string;
  };
  // ... other fields you need
};
export interface EventDuplicatePayload extends EventBase {
    duplicate_products: boolean;
    duplicate_questions: boolean;
    duplicate_settings: boolean;
    duplicate_promo_codes: boolean;
    duplicate_capacity_assignments: boolean;
    duplicate_check_in_lists: boolean;
    duplicate_event_cover_image: boolean;
    duplicate_webhooks: boolean;
}
// ...
export interface SortDirectionLabel {
    asc: string;
    desc: string;
}

export interface PaginationData {
    total: number;
    per_page: number;
    current_page: number;
    links: string[];
    last_page: number;
    from: number;
    to: number;
    path: number;
    allowed_sorts: Record<string, SortDirectionLabel>;
    default_sort: string;
    default_sort_direction: string;
}

export interface GenericDataResponse<T> {
    data: T;
    errors?: Record<string, string>;
}

export interface GenericPaginatedResponse<T> {
    data: T[];
    meta: PaginationData;
}
// ...
export type QueryFilterCondition = {
    operator: QueryFilterOperator;
    value: QueryFilterValue;
};

export type QueryFilterFields = {
    [key: string]: QueryFilterCondition | QueryFilterCondition[] | undefined;
}

export interface QueryFilters {
    pageNumber?: number;
    perPage?: number;
    query?: string;
    sortBy?: string;
    sortDirection?: string;
    filterFields?: QueryFilterFields;
    additionalParams?: Record<string, any>;
}